[engraving] replaced QList to std. Part 4

+++ src/engraving/libmscore/bracketItem.cpp
@@ -51,7 +51,7 @@ mu::engraving::PropertyValue BracketItem::getProperty(Pid id) const
     case Pid::BRACKET_COLUMN:
         return _column;
     case Pid::BRACKET_SPAN:
-        return _bracketSpan;
+        return static_cast<int>(_bracketSpan);
     default:
         return EngravingItem::getProperty(id);
     }
@@ -67,7 +67,7 @@ bool BracketItem::setProperty(Pid id, const engraving::PropertyValue& v)
         staff()->changeBracketColumn(column(), v.toInt());
         break;
     case Pid::BRACKET_SPAN:
-        _bracketSpan = v.toInt();
+        _bracketSpan = static_cast<size_t>(v.toInt());
         break;
     default:
         return EngravingItem::setProperty(id, v);
