+++ src/engraving/libmscore/rootitem.cpp
@@ -56,7 +56,7 @@ EngravingObject* RootItem::scanParent() const
     return m_score->scanParent();
 }
 
-EngravingObject* RootItem::scanChild(int n) const
+EngravingObject* RootItem::scanChild(size_t n) const
 {
     return m_score->scanChild(n);
 }
